refactor tests into appropriately named files
